Exprimer les niveaux de tuile directement en mètres

+++ tilecoveragecalculator.h
@@ -9,7 +9,7 @@
 class TileCoverageCalculator
 {
 public:
-    static QVector<Tuile> rectangle(const EntityBounds &bounds, int level);
+    static QVector<Tuile> rectangle(const EntityBounds &bounds, qint64 level);
     static qint64 floorToGrid(double coordinate, qint64 tileSize);
 };
